+++ @types/index.d.ts
@@ -6,3 +6,17 @@ declare module 'consolidate';
 interface ObjectValueType {
   [key: string]: string | boolean | { [key: string]: any }
 }
+
+interface MacrosType {
+  scriptName: string,
+  placeholder: string,
+  download: string,
+  remaxRepo: string,
+  templateRepo: string,
+  templateTSRepo: string,
+  description: string,
+  descriptionTS: string,
+  tmpPathName: string,
+  templatePathName: string,
+  defaultPlatform: string
+}
