+++ src/utils/utils.ts
@@ -1,0 +1,5 @@
+export const firstUpperCase = (str: string) => {
+  return str.toLowerCase().replace(/^./, (s: string): string => {
+    return s.toUpperCase()
+  })
+}
